Update & Bug fix:
- code has been slightly updated fixing a couple of issues that came out
from compilation on a system different from that of the original development:
 > SOLVE_GZ module wrapper has been introduced. This module wrap all the GZ
code modules and is or should be the only "exposed" to the user --> going
towards an opaque code!! (preserve the memory)
 > LANCELOT_WRAP.f90 has been introduced. This module provides an interface to
the lancelot_solve procedure in Lancelot B. We got rid of probfun procedure as
the function to be minimzed is specified direclty in the call to
lancelot_simple.
 > update of gzproj_min_routines.f90 and slater_min_routines.f90 containing
a call to fsolve with function procedure interface. In the latest
SCIFOR release this procedure is now distinct from fsolve
(subroutine interface) and it is called  f_fsolve. This is because some
GNU compiler apparently complain about the overload of the fsolve_func and
fsolve_sub procedure in SF_OPTIMIZE.f90
 > introduced make.inc and cleaned out the Makefile (!!)
 > Updated GZ_ENERGY and GZ_OPTIMIZE to point to LANCELOT_WRAP rather
than lancelot_simple_double.