Changes in the determination of the optimized variational matrices:
Possibility to read the full matrices (be aware that in some cases the outputs
that were produced in the equilibrium calculations were wrong) and then find
the basis decomposition.

This is useful for SC calculations starting from the normal state.

Track of the weight of the 210 configutration for the k3c60 driver.